Release/0.2.0

+++ lib/UART-Arduino/UART-Arduino.cpp
@@ -7,19 +7,19 @@ UART_Arduino::UART_Arduino(HardwareSerial& serial)
     _serial.setTimeout(_timeout);
 }
 
-bool UART_Arduino::begin(uint64_t baud)
+bool UART_Arduino::begin(uint32_t baud)
 {
     _serial.begin(baud);
     return true;
 }
 
-void UART_Arduino::setTimeout(uint64_t timeout)
+void UART_Arduino::setTimeout(uint32_t timeout)
 {
     UART_Interface::setTimeout(timeout);
     _serial.setTimeout(timeout);
 }
 
-uint64_t UART_Arduino::getTimeout(void) const
+uint32_t UART_Arduino::getTimeout(void) const
 {
     return UART_Interface::getTimeout();
 }
